grafana/desc_editor/requirements.txt
streamlit>=1.39.0
requests>=2.32.0
python-dotenv>=1.0.1
pandas>=2.2.0
plotly>=5.18.0
